calculates eigenvector centrality and the two measures of extended niche.

calculates eigenvector centrality and the two measures of extended niche.

Usage

node_topol(int_data, cover_data, int_type = c("rec", "fac", "comp"))

Arguments

int_data

data frame containing interaction data.

cover_data

data frame with the abundance of each canopy species in each plot.

int_type

Indicates the type of plant-plant interaction that will be presented in the output: general recruitment, recruitment enhancement (i.e. facilitation) or recruitment depression (i.e. competition).

rec: Estimates eigenvector centrality and extended neighborhoods for the general recruitment network.

fac: Estimates eigenvector centrality and extended neighborhoods for the facilitation network.

comp: Estimates eigenvector centrality and extended neighborhoods for the competition network.

Value

data frame

a data frame with estimates of the implications of each plant species in indirect interactions with the rest of species, both as canopy and recruit. Specifically, the estimates provided are the following: For recruitment networks:

Eigenvector centrality: provides the dominant eigenvector of the adjacency matrix of the interactions network. Its entries are interpreted as the centrality of each node. A node is more influential if it is connected to other influential nodes. In the case of recruitment networks, it has been associated with the persistence of species (Alcantara and Rey 2012; Alcantara et al. 2017): species with 0 eigenvector centrality have the lowest probability of persisting in the local community.

Extended canopy service: The extended canopy service of a species is defined as the number of other species whose recruitment it enables, directly or indirectly. It reflects an expansion of the species' out-degree in the interaction network, capturing its broader influence on the recruitment dynamics within the community. Using the logic of replacement dynamics, through a temporal series of replacements, a spatial patch will be dominated successively by plants of different species. The legacies of each species on the soil may remain for long time, so the occupation of a patch by one species can affect the recruitment of many others. From a functional perspective, it can be interpreted in terms of how easily disturbances associated with one species may pervade to the rest of species directly or indirectly; for example, a pest decimating the population of one species would affect many or few other species depending on the size of its extended canopy service.

Extended recruitment niche: The number of species in the community whose presence may allow eventually (directly and indirectly) the recruitment of the focal species. It represents an extension of the species' in-degree, capturing its dependence on the broader network for successful recruitment. From a functional perspective, a species with a large extended recruitment niche will be more susceptible to pests affecting any species of the community, but at the same time it may benefit from the increase in abundance of any species in its neighborhood. For facilitation networks:

Eigenvector centrality: Provides the dominant eigenvector of the adjacency matrix of the interactions network, and its entries are interpreted as the centralities of the nodes. It provides a relative value (i.e., proportionality) indicating for each nurse plant, the extent to which it tends to facilitate—directly or indirectly—those species that in turn facilitate many others. The centrality of a species is proportional to the sum of the centralities of the species it facilitates.

Extended nurse service: The number of species that a given species facilitates, either directly or indirectly (i.e., the number of nodes that can be reached from a given node).

Extended facilitated niche: The number of nurse species that either directly promote the recruitment of a given species or indirectly facilitate the recruitment of its own nurse plants (i.e. the number of nodes that can reach this node). A high value indicates that this species recruits under nurse plants that themselves recruit under other nurse plants. This metric can reflect how sensitive a species is to the loss of nurse plants, especially when the ratio between the number of nurse species (i.e., degree) and the extended recruitment niche is low. For competition networks:

Eigenvector centrality: Provides a relative value (i.e., proportionality) indicating for each canopy species, the extent to which it tends to depress the recruitment of those species that, in turn, depresses many others when they act as canopy. The centrality of a species is proportional to the sum of the centralities of the species it negatively affects.

Extended competitor effect: The number of species whose recruitment is depressed by a given canopy species, either directly or indirectly depressing its depressors (i.e., the number of nodes that can be reached from a given node). The odd or even number of links in the chain can lead to different outcomes, with even-numbered chains potentially canceling out some effects, while odd-numbered chains can amplify them.

Extended competitors niche: A high value indicates that the recruitment of this species is depressed by species whose recruitment is depressed by other recruitment-depressing canopy species. It represents the number of recruitment-depressing canopy species that either directly depress its recruitment, or indirectly affect the recruitment of its depressor canopy species (i.e., the number of nodes that can reach this node). The odd or even number of links in the chain can lead to different outcomes, with even-numbered chains potentially canceling out some effects, while odd-numbered chains can amplify them. This metric can reflect the extent to which a species can escape its depressors due to the presence of other species that suppress the recruitment of its depressors, especially when the ratio between the number of depressing canopy species (i.e., degree) and the extended depressed niche is low.

Examples

summary_node_topol_rec <- node_topol(Amoladeras_int,Amoladeras_cover, int_type="rec")
head(summary_node_topol_rec)
#>                      Eigenvector centrality Extended canopy service
#> Artemisia_barrelieri                 0.3840                      20
#> Artemisia_campestris                 0.1083                       0
#> Asparagus_albus                      0.2638                      20
#> Asparagus_horridus                   0.6316                      20
#> Ballota_hirsuta                      0.4896                      20
#> Helichrysum_stoechas                 0.6499                      20
#>                      Extended recruitment niche
#> Artemisia_barrelieri                         20
#> Artemisia_campestris                         21
#> Asparagus_albus                              20
#> Asparagus_horridus                           20
#> Ballota_hirsuta                              20
#> Helichrysum_stoechas                         20

summary_node_topol_fac <- node_topol(Amoladeras_int,Amoladeras_cover, int_type="fac")
head(summary_node_topol_fac)
#>                      Eigenvector nurse centrality Extended nurse service
#> Artemisia_barrelieri                       0.3921                     18
#> Artemisia_campestris                       0.0000                      0
#> Asparagus_albus                            0.1281                     18
#> Asparagus_horridus                         0.1506                     18
#> Ballota_hirsuta                            0.2645                     18
#> Helichrysum_stoechas                       0.5314                     18
#>                      Extended facilitated niche
#> Artemisia_barrelieri                         17
#> Artemisia_campestris                         18
#> Asparagus_albus                              17
#> Asparagus_horridus                           17
#> Ballota_hirsuta                              17
#> Helichrysum_stoechas                         17

summary_node_topol_comp <- node_topol(Amoladeras_int,Amoladeras_cover, int_type="comp")
head(summary_node_topol_comp)
#>                      Eigenvector canopy centrality
#> Artemisia_barrelieri                             0
#> Helichrysum_stoechas                             0
#> Ononis_natrix                                    1
#> Phagnalon_saxatile                               0
#> Teucrium_polium                                  0
#> Thymelaea_hirsuta                                0
#>                      Extended canopy depression effect
#> Artemisia_barrelieri                                 0
#> Helichrysum_stoechas                                 0
#> Ononis_natrix                                        5
#> Phagnalon_saxatile                                   1
#> Teucrium_polium                                      0
#> Thymelaea_hirsuta                                    0
#>                      Extended recruitment depression 
#> Artemisia_barrelieri                                2
#> Helichrysum_stoechas                               13
#> Ononis_natrix                                      11
#> Phagnalon_saxatile                                 12
#> Teucrium_polium                                    12
#> Thymelaea_hirsuta                                  12