Link_completeness

Observed and estimated number of links, and link completeness. Methods based in Chao et al. (2014) as implemented in R package iNEXT.

link_completeness(
  dataset = NULL,
  site = NULL,
  type = c("incidence", "abundance")
)

Arguments

dataset

The name of the data set where recruitment networks are stored. Usually, it will be "RecruitNet" (or the name assigned to the data set when it was imported). At least, the data set must contain columns named "Study_site", "Plot", "Canopy", "Recruit" and "Frequency". In data sets containing a single site or a single plot, the corresponding columns must be included anyway.

site

The name of a study site.

type

Can take two values: "incidence" or "abundance". Estimates based on incidence data are recommended, but can only be obtained if your data set is structured in multiple plots. Alternatively, for data collected in a single plot, estimates can be based only on abundance data.

Value

A data frame TODO

Examples

if (FALSE) { # interactive()

link_completeness(RecruitNet, "Ventisquero", "incidence")
link_completeness(RecruitNet, "Laxe", "abundance")
link_completeness(RecruitNet, "Laxe", "incidence") # Issues a warning
link_completeness(RecruitNet, "LosReyes", "abundance")
link_completeness(RecruitNet, "LosReyes", "incidence") #Issues an error
}